packaging started

+++ packagename/wrapped.pyx
@@ -2,7 +2,7 @@ cimport cython
 import numpy as np
 cimport numpy as np
 
-DTYPE = np.float
+DTYPE = float
 ctypedef np.float_t DTYPE_t
 
 # This file shows 4 examples:
